Move low-level cache tests to generic file
Also round out some of the tests by adding a bit more, especially for
the new functions.
